Add OpenGraph metadata attributes to index.html
The image presents a diagram in English, and the messages
are bilingual, which is a little unfortunate. It would have
been better to have multiple index.htmls, one per language,
and have the app automatically replace the URL in the location
bar as needed; but that introduces additional complexity, so
this’ll do for now.